feat: 日程插件适配
日程插件适配

Log: 日程插件适配

Task: https://pms.uniontech.com/task-view-132599.html

+++ calendar-service/src/dbmanager/daccountdatabase.h
@@ -53,7 +53,7 @@ public:
     //根据关键字查询一定范围内的日程
     DSchedule::List querySchedulesByKey(const QString &key);
     //根据重复规则查询一定范围内的日程
-    DSchedule::List querySchedulesByRRule(const QString &key, const ushort &rruleType);
+    DSchedule::List querySchedulesByRRule(const QString &key, const int &rruleType);
     //获取需要提醒的日程信息
     DSchedule::List getRemindSchedule();
